Various small things

+++ src/main/java/com/github/antilaby/antilaby/update/Updater.java
@@ -15,11 +15,9 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.net.URL;
-import java.net.URLConnection;
 
 /**
- * The old auto-updater class. It will be removed as soon as the new
- * auto-updater is ready for use.
+ * The old auto-updater class. It will be removed as soon as the new auto-updater is ready for use.
  *
  * @author NathanNr
  */
@@ -33,38 +31,37 @@ public class Updater extends Thread {
 		AntiLaby.LOG.debug("Checking for updates on spigotmc.org...");
 		String newVersion;
 		try {
-			newVersion = IOUtils
-					.readUrl("https://api.spigotmc.org/legacy/update.php?resource=" + Constants.RESOURCE_ID);
-			if (!newVersion.isEmpty()) {
-				if (!newVersion.contains(" ") || !newVersion.contains("!")) {
-					if (!newVersion.equalsIgnoreCase(AntiLaby.getInstance().getDescription().getVersion())) {
+			newVersion = IOUtils.readUrl(Constants.UPDATE_URL);
+			if(!newVersion.isEmpty()) {
+				if(!newVersion.contains(" ") || !newVersion.contains("!")) {
+					if(!newVersion.equalsIgnoreCase(AntiLaby.getInstance().getDescription().getVersion())) {
 						AntiLaby.LOG.info("Update found! Version " + newVersion + " is available.");
 					} else {
 						AntiLaby.LOG.debug("No update found. You are running the newest version.");
 						return;
 					}
 				} else {
-					AntiLaby.LOG.error("Failed to check for updates on spigotmc.org! (Invalid value received: "
-							+ newVersion + ")");
+					AntiLaby.LOG.error(
+							"Failed to check for updates on spigotmc.org! (Invalid value received: " + newVersion +
+									")");
 					return;
 				}
 			} else {
 				AntiLaby.LOG.error("Failed to check for updates on spigotmc.org! (No information received)");
 				return;
 			}
-		} catch (final Exception ex) {
+		} catch(final Exception ex) {
 			AntiLaby.LOG.error("Failed to check for updates on spigotmc.org! (" + ex.getMessage() + ")");
 			return;
 		}
 		// Download and install update if available
-		if (installUpdate("https://github.com/NathanNr/AntiLaby/releases/download/" + newVersion + "/AntiLaby.jar"))
+		if(installUpdate("https://github.com/NathanNr/AntiLaby/releases/download/" + newVersion + "/AntiLaby.jar"))
 			AntiLaby.LOG.info("Auto-update complete! Reload or restart your server to activate the new version.");
-		else
-			AntiLaby.LOG.error("Failed to install update! Please install the newest version manually from "
-					+ Constants.RESOURCE_LINK + "/!");
+		else AntiLaby.LOG.error(
+				"Failed to install update! Please install the newest version manually from " + Constants.RESOURCE_LINK
+						+ "!");
 		final File tmp = new File("plugins/AntiLaby.tmp");
-		if (tmp.exists())
-			tmp.delete();
+		if(tmp.exists()) tmp.delete();
 		AntiLaby.getInstance().disableIfNotCompatible();
 		interrupt();
 	}
@@ -73,44 +70,36 @@ public class Updater extends Thread {
 	 * Tries to download and install the update if enabled in the config
 	 *
 	 * @param urlString
-	 *            the Non-null update URL
+	 * 		the Non-null update URL
 	 *
 	 * @return true if the update task completed, false otherwise
 	 */
 	private boolean installUpdate(@Nonnull String urlString) {
 		AntiLaby.LOG.info("Downloading update from download server...");
 		try {
-			final URL url = new URL(urlString);
-			final URLConnection conn = url.openConnection();
-			final InputStream is = new BufferedInputStream(conn.getInputStream());
+			final InputStream is = new BufferedInputStream(new URL(urlString).openConnection().getInputStream());
 			final OutputStream os = new BufferedOutputStream(new FileOutputStream("plugins/AntiLaby.tmp"));
-			final byte[] chunk = new byte[1024];
-			int chunkSize;
-			while ((chunkSize = is.read(chunk)) != -1)
-				os.write(chunk, 0, chunkSize);
+			IOUtils.copyStream(is, os);
 			os.close();
-			final File newfile = new File("plugins/AntiLaby.tmp");
-			final long newlength = newfile.length();
-			if (newlength <= 10000) {
-				newfile.delete();
+			final File tempFile = new File("plugins/AntiLaby.tmp");
+			final long newlength = tempFile.length();
+			if(newlength < 10000) {
+				tempFile.delete();
 				return false;
 			} else {
 				AntiLaby.LOG.info("Installing update...");
-				final FileInputStream is2 = new FileInputStream(newfile);
+				final FileInputStream is2 = new FileInputStream(tempFile);
 				final OutputStream os2 = new BufferedOutputStream(
 						new FileOutputStream(AntiLaby.getInstance().getFile()));
-				final byte[] chunk2 = new byte[1024];
-				int chunkSize2;
-				while ((chunkSize2 = is2.read(chunk2)) != -1)
-					os2.write(chunk2, 0, chunkSize2);
+				IOUtils.copyStream(is2, os2);
 				is2.close();
 				os2.close();
-				newfile.delete();
+				tempFile.delete();
 				return true;
 			}
-		} catch (final FileNotFoundException e) {
+		} catch(final FileNotFoundException e) {
 			return false;
-		} catch (final IOException e) {
+		} catch(final IOException e) {
 			e.printStackTrace();
 			return false;
 		}
